Basics of Dog class
This area has to be after one blank line, the first line being the
summary of an idea.  This area being 80 characters wide, and as long as
ou want it to be.  But describing ideas, not specific code.

The current changes can be described as this (and when your future self
reads this you will shed a tear of joy for the understanding that you
have chosen to impart at this time):

Fixes the problems associated with having output in the methods, now
bark and other methods return a string which can be used as the user
wants, trying to keep unnecessary opinions of ours out of it.

There is a story used in a HereDoc format that demonstrates what happens
when puts is used in the method.

Also introduced is the "application block" which does not run when the
file is loaded as a library (which includes a testing framework) but
does run when ran directly.